Modify/add index.html and style.css
Added to the main section. Created two new sections nested within the main. One section was for a gallery, the other for a quote. Then added the call to action section at the bottom. The call to action is the last piece of the main section. It also contains a button.